LayoutTests/svg/dynamic-updates/SVGFilterElement-svgdom-primitiveUnits-prop-expected.txt

CONSOLE MESSAGE: line 48: TypeError: 'undefined' is not an object (evaluating 'filterElement.primitiveUnits.baseVal = SVGUnitTypes.SVG_UNIT_TYPE_OBJECTBOUNDINGBOX')
FAIL: Timed out waiting for notifyDone to be called
SVG 1.1 dynamic update tests

Tests dynamic updates of the 'primitiveUnits' property of the SVGFilterElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL filterElement.primitiveUnits.baseVal should be 1. Threw exception TypeError: 'undefined' is not an object (evaluating 'filterElement.primitiveUnits.baseVal')
